backend/apps/usuarios/views.py
from django.shortcuts import render, get_object_or_404
from rest_framework import status
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated
from datetime import date

from .serializer import (
    UserSerializer,
)

from .models import (
    User,
)
